Start a coverage module
