+++ tooling/nargo_cli/src/cli/test_cmd.rs
@@ -24,14 +24,14 @@ use nargo::{
     errors::CompileError,
     foreign_calls::DefaultForeignCallBuilder,
     insert_all_files_for_workspace_into_file_manager,
-    ops::{FuzzConfig, TestStatus, check_crate_and_report_errors, report_errors},
+    ops::{FuzzConfig, TestStatus, report_errors},
     package::Package,
     parse_all, prepare_package,
     workspace::Workspace,
 };
 use nargo_toml::PackageSelection;
 use noirc_driver::{CompileOptions, check_crate};
-use noirc_errors::{Location, reporter::ReportedErrors};
+use noirc_errors::reporter::ReportedErrors;
 use noirc_frontend::graph::CrateId;
 use noirc_frontend::hir::{
     Context, FunctionNameMatch, ParsedFiles, comptime::EvaluationTracker, def_map::TestFunction,
@@ -41,6 +41,7 @@ use crate::errors::CliError;
 
 use super::{LockType, PackageOptions, WorkspaceCommand};
 
+mod coverage;
 pub(crate) mod formatters;
 
 /// Fully qualified test name.
@@ -784,27 +785,4 @@ impl<'a> TestRunner<'a> {
             self.args.compile_options.silence_warnings,
         )
     }
-}
-
-/// Returns the location of every expression in the crate that is not inside a `#[test]` function.
-/// Used to build the zero-count baseline for lcov coverage reports.
-fn baseline_expr_locations(context: &Context, crate_id: CrateId) -> Vec<Location> {
-    let def_map = &context.def_maps[&crate_id];
-    let allowed_files = def_map.file_ids();
-
-    let test_body_spans: Vec<Location> = def_map
-        .get_all_test_functions(&context.def_interner)
-        .filter_map(|test_fn| context.def_interner.function(&test_fn.id).try_as_expr())
-        .map(|body_id| context.def_interner.expr_location(&body_id))
-        .collect();
-
-    context
-        .def_interner
-        .expr_locations_for_files(&allowed_files)
-        .filter(|loc| {
-            !test_body_spans
-                .iter()
-                .any(|test_loc| test_loc.file == loc.file && test_loc.span.contains(&loc.span))
-        })
-        .collect()
 }
